Move to lazy references instead of copies
Instead of copying things around when we do select/map,
this new version now relies on a pair of source iterators.

No work is done until the results are evaluated:

    std::vector<int> vec = { 1, 2, 3, 4 };
    auto enumerable = Enumerable(vec.begin(), vec.end());
    auto filtered = enumerable.select([](auto&& i) { return i % 2 == 0 });

    for (auto&& v : filtered) { std::cout << v << std::endl; }

The actual filtering only happens when we're iterating, not when
filtered is declared.

This means that the source objects must exist while the enumerables are
alive, otherwise we'll be accessing freed memory.